requirements.txt
aws-cdk-lib==2.29.0
aws-cdk.aws-glue-alpha==2.29.0a0
aws-cdk.aws-kinesisfirehose-alpha==2.29.0a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.29.0a0
constructs>=10.0.0,<11.0.0
aws_ddk_core==0.5.1
boto3==1.21.4
wrapt==1.14.1
cdk-watchful==0.6.256